Remove auth
Embedding auth credentials in the URL broke the breadcrumbs which
are useful for navigation and I did not want to make the user
type in credentials manually. Making it work automatically without
baking it into the URL would involve some fragile hacks.
Given that this goes through a tunneled port-forward anyway and
at least uses a random port it doesn't materially affect security.

An automatic proxy for this will be nice.